fix some bugs of delete

+++ include/WhereState.h
@@ -6,7 +6,7 @@
 
 typedef struct SelectCols {
     int *idxList;
-    size_t idxListLen;
+    int idxListLen;
 } SelectCols_t;
